modified to get station description data for interesting months
modified:   bdp.pl
hacked in another regex to get station text files, and another loop
to do the downloading.  Only download if there is something interesting
in the data for that month.  Doesn't necessarily mean that the definition
file is actually dated similar.  For example, in d6, 2007 data is downloaded, but
themost recent station def file is dated 2005.