fixed case where speed directions and wim directions do not overlap
Presuming that speed data is better than the raw wim data
 modified:   Rwork/wim.pre.processing.R

This is a strange thing.  I have a problem at site 10, in Fresno.  The whole way the WIM
sites are labeled by lane and direction is flawed.  I need to get on the phone and
yell at Caltrans to get their act together.