working on code to parameterize count variables
A question was asked about what about empty trucks.
I can't answer that now, so I thought it would be a good
exercise and addition to the project to be able to add that
So I need to parameterize the count variables, and add hooks or
something for the sql query code.

 new file:   Rwork/wim.cleanup.lane.assignments.R
 modified:   Rwork/wim.pre.processing.R
 modified:   Rwork/wim.vds.processing.functions.R