modified impute_to_couchdb to save many docs per imputation, and combine lanes
other way choked on really big document (after multiple imputation)
Trying this way to avoid the crashing...many small documents, one
document per hour imputed.   Hmm, if a value is not an imputed value, then
it will be saved multiple times for no reason, so should check that as well...
forumulate the doc id based on some hashing of the content, so that
conflicts will be obvious.