a great optimization idea turns out not so
had an idea to keep file handles around as I roll through files.

didn't work out.  Writing not that slow I guess, plus it is a pain to
keep the file handles open...would be difficult for others to use

Instead, I keep track of filenames, saving a little bit of time (like
the call to -e everytime I write to a file.

test passes, so at least no regressions!